+++ src/client/java/tytoo/weave/animation/PropertyInterpolator.java
@@ -2,5 +2,5 @@ package tytoo.weave.animation;
 
 @FunctionalInterface
 public interface PropertyInterpolator<T> {
-    T interpolate(T start, T end, float progress);
+    T interpolate(T start, T end, double progress);
 }
